Image Customizer: Make grub config customization more robust.
This change adds a tokenizer for grub config files. While not a full
parser, it does ensure that semicolons, character escapes,
single-quotes, and double-quotes are all parsed properly. Using this
tokenizer will allow the image customizer to handle a greater variety
of grub config files.

In addition, all existing image customizer code that modifies the grub
config file has been updated to use the grub tokenizer instead of using
regexes and the like.